impl(shaped): Support import to PC sheet w/ 5e Shaped Sheet.
This is a very quick-and-dirty implementation.
For now it breaks existing OGL Sheet support in several places.

Supported functionality for 5e Shaped Sheet is equivalent to or slightly
better than previous support for OGL Sheet (in scope of imported
features and data granularity). The only exception is Psionics - they
are not supported yet.